requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
blobfile==2.1.1
    # via tiktoken
brotli==1.1.0
    # via yt-dlp
certifi==2024.7.4
    # via
    #   httpx
    #   requests
    #   yt-dlp
chardet==3.0.4
    # via
    #   httpx
    #   requests
clyent==1.2.1
    # via -r requirements.in
filelock==3.15.4
    # via
    #   blobfile
    #   torch
fsspec==2024.6.1
    # via torch
googletrans==4.0.0rc1
    # via -r requirements.in
h11==0.9.0
    # via httpcore
h2==3.2.0
    # via httpcore
hpack==3.0.0
    # via h2
hstspreload==2024.8.1
    # via httpx
httpcore==0.9.1
    # via httpx
httpx==0.13.3
    # via
    #   -r requirements.in
    #   googletrans
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   -r requirements.in
    #   httpx
    #   requests
jinja2==3.1.4
    # via torch
llvmlite==0.43.0
    # via numba
lxml==4.9.4
    # via blobfile
markupsafe==2.1.5
    # via jinja2
more-itertools==10.3.0
    # via openai-whisper
mpmath==1.3.0
    # via sympy
mutagen==1.47.0
    # via yt-dlp
networkx==3.3
    # via torch
numba==0.60.0
    # via openai-whisper
numpy==2.0.1
    # via
    #   numba
    #   openai-whisper
openai-whisper==20231117
    # via -r requirements.in
pycryptodomex==3.20.0
    # via
    #   blobfile
    #   yt-dlp
pydub==0.25.1
    # via -r requirements.in
regex==2024.7.24
    # via tiktoken
requests==2.25.1
    # via
    #   -r requirements.in
    #   requests-mock
requests-mock==1.12.1
    # via -r requirements.in
rfc3986==1.5.0
    # via httpx
sniffio==1.3.1
    # via
    #   httpcore
    #   httpx
sympy==1.13.1
    # via torch
tiktoken==0.1.1
    # via openai-whisper
torch==2.4.0
    # via openai-whisper
tqdm==4.66.5
    # via openai-whisper
typing-extensions==4.12.2
    # via torch
urllib3==1.26.19
    # via
    #   blobfile
    #   requests
websockets==12.0
    # via yt-dlp
yt-dlp==2023.10.13
    # via -r requirements.in
